+++ src/Patch.cpp
@@ -132,7 +132,6 @@ int xsum,ysum;
 short hx;
 float k,q,envval;
 
-//pCell = NULL; 
 localK = 0.0; // no. of suitable cells (unadjusted K > 0) in the patch
 int nsuitable = 0;
 double mean;
@@ -176,7 +175,7 @@ for (int i = 0; i < ncells; i++) {
 	switch (rasterType) {
 	case 0: // habitat codes
 		hx = cells[i]->getHabIndex(landIx);
-		k = (float)pSpecies->getHabK(hx);
+		k = pSpecies->getHabK(hx);
 		if (k > 0.0) {
 			nsuitable++;
 			localK += envval * k;
@@ -186,7 +185,7 @@ for (int i = 0; i < ncells; i++) {
 		k = 0.0;
 		for (int j = 0; j < nHab; j++) { // loop through cover layers
 			q = cells[i]->getHabitat(j);
-			k += q * (float)pSpecies->getHabK(j) / 100.0f;
+			k += q * pSpecies->getHabK(j) / 100.0f;
 		}
 		if (k > 0.0) {
 			nsuitable++;
@@ -197,7 +196,7 @@ for (int i = 0; i < ncells; i++) {
 		q = cells[i]->getHabitat(landIx);
 		if (q > 0.0) {
 			nsuitable++;
-			localK += envval * (float)pSpecies->getHabK(0) * q / 100.0f;
+			localK += envval * pSpecies->getHabK(0) * q / 100.0f;
 		}
 		break;
 	}
@@ -225,14 +224,14 @@ if (env.stoch && env.inK) { // environmental stochasticity in K
 	// apply min and max limits to K over the whole patch
 	// NB limits have been stored as N/cell rather than N/ha
 	float limit;
-	limit = (float)pSpecies->getMinMax(0) * (float)nsuitable;
+	limit = pSpecies->getMinMax(0) * (float)nsuitable;
 	if (localK < limit) localK = limit;
 #if RSDEBUG
 //DEBUGLOG << "Patch::setCarryingCapacity(): patchNum=" << patchNum
 //	<< " limit=" << limit << " localK=" << localK
 //	<< endl;
 #endif
-	limit = (float)pSpecies->getMinMax(1) * (float)nsuitable;
+	limit = pSpecies->getMinMax(1) * (float)nsuitable;
 	if (localK > limit) localK = limit;
 #if RSDEBUG
 //DEBUGLOG << "Patch::setCarryingCapacity(): patchNum=" << patchNum
